+++ CHANGELOG.md
@@ -2,19 +2,29 @@
 
 All notable changes to the "flutterflow-code" extension will be documented in this file.
 
+## [v0.0.8]
+
+Added support for Enterprise India
+
 ## [v0.0.7]
+
 Fixed a pescky path bug that recursively copied code
 Some code refactoring
+
 ## [v0.0.5]
+
 Fixed a folder rename bug
+
 ## [v0.0.4]
 
 - Added support for windows devices
-- Added support to initiate a flutter run session 
+- Added support to initiate a flutter run session
 - Added support to initialize git locally
+
 ## [v0.0.3]
 
 - Added support for code download without assets
+
 ## [v0.0.2]
 
 - Initial release